requirements.txt
celery==5.4.0
Django==5.1.3
django-celery-beat==2.7.0
ipython==8.29.0
redis==5.2.0
python-dotenv==1.0.1
django-apscheduler==0.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
django-cors-headers==4.6.0
flake8==7.1.1
drf-yasg==1.21.8
requests==2.32.3
eventlet==0.38.0
pytest==8.3.3
coverage==7.6.8
pytest-django==4.9.0
